feat(repo): switch to using dependsOn for the relationship between build and build-base
