+++ e2e/angular/src/angular-core.test.ts
@@ -1,3 +1,5 @@
+process.env.SELECTED_CLI = 'angular';
+
 import * as path from 'path';
 import {
   checkFilesExist,
@@ -12,6 +14,7 @@ import {
   uniq,
   updateFile,
 } from '@nrwl/e2e/utils';
+
 import { names } from '@nrwl/devkit';
 
 describe('Angular Package', () => {
@@ -146,6 +149,7 @@ describe('Angular Package', () => {
 
   it('should support eslint and successfully lint external HTML files and inline templates', async () => {
     const myapp = uniq('myapp');
+
     runCLI(`generate @nrwl/angular:app ${myapp} --linter=eslint`);
 
     const templateWhichFailsBananaInBoxLintCheck = `<div ([foo])="bar"></div>`;
@@ -174,58 +178,17 @@ describe('Angular Package', () => {
     );
 
     const appLintStdOut = runCLI(`lint ${myapp}`, { silenceError: true });
-
     expect(appLintStdOut).toContain(
       path.normalize(`apps/${myapp}/src/app/app.component.html`)
     );
-    expect(appLintStdOut).toContain(
-      `1:6  error  Invalid binding syntax. Use [(expr)] instead  @angular-eslint/template/banana-in-box`
-    );
+    expect(appLintStdOut).toContain(`1:6`);
+    expect(appLintStdOut).toContain(`Invalid binding syntax`);
     expect(appLintStdOut).toContain(
       path.normalize(`apps/${myapp}/src/app/inline-template.component.ts`)
     );
     expect(appLintStdOut).toContain(
-      `5:21  error  The selector should be prefixed by one of the prefixes: '${proj}' (https://angular.io/guide/styleguide#style-02-07)  @angular-eslint/component-selector`
+      `The selector should be prefixed by one of the prefixes`
     );
-    expect(appLintStdOut).toContain(
-      `7:18  error  Invalid binding syntax. Use [(expr)] instead                                                                  ${' '.repeat(
-        proj.length
-      )}@angular-eslint/template/banana-in-box`
-    );
-
-    const mylib = uniq('mylib');
-    runCLI(`generate @nrwl/angular:lib ${mylib} --linter=eslint`);
-
-    // External HTML template file
-    updateFile(
-      `libs/${mylib}/src/lib/some.component.html`,
-      templateWhichFailsBananaInBoxLintCheck
-    );
-
-    // Inline template within component.ts file
-    updateFile(
-      `libs/${mylib}/src/lib/inline-template.component.ts`,
-      wrappedAsInlineTemplate
-    );
-
-    const libLintStdOut = runCLI(`lint ${mylib}`, { silenceError: true });
-
-    expect(libLintStdOut).toContain(
-      path.normalize(`libs/${mylib}/src/lib/some.component.html`)
-    );
-    expect(libLintStdOut).toContain(
-      `1:6  error  Invalid binding syntax. Use [(expr)] instead  @angular-eslint/template/banana-in-box`
-    );
-    expect(libLintStdOut).toContain(
-      path.normalize(`libs/${mylib}/src/lib/inline-template.component.ts`)
-    );
-    expect(libLintStdOut).toContain(
-      `5:21  error  The selector should be prefixed by one of the prefixes: '${proj}' (https://angular.io/guide/styleguide#style-02-07)  @angular-eslint/component-selector`
-    );
-    expect(libLintStdOut).toContain(
-      `7:18  error  Invalid binding syntax. Use [(expr)] instead                                                                  ${' '.repeat(
-        proj.length
-      )}@angular-eslint/template/banana-in-box`
-    );
+    expect(appLintStdOut).toContain(`7:18`);
   });
 });
