+++ e2e/angular/src/config-compat.test.ts
@@ -1,3 +1,5 @@
+process.env.SELECTED_CLI = 'angular';
+
 import {
   expectTestsPass,
   newProject,
