+++ e2e/angular/src/ng-add.test.ts
@@ -1,3 +1,5 @@
+process.env.SELECTED_CLI = 'angular';
+
 import {
   checkFilesExist,
   readJson,
