+++ jest.preset.js
@@ -1,5 +1,7 @@
 const nxPreset = require('@nrwl/jest/preset');
 
+process.env.npm_config_registry = `http://localhost:4872`;
+
 module.exports = {
   ...nxPreset,
   testTimeout: 30000,
