+++ nx.json
@@ -1,7 +1,6 @@
 {
   "npmScope": "nrwl",
   "implicitDependencies": {
-    "tslint.json": "*",
     ".eslintrc.json": "*"
   },
   "tasksRunnerOptions": {
@@ -20,160 +19,116 @@
       }
     }
   },
+  "targetDependencies": {
+    "build": [
+      {
+        "target": "build-base",
+        "projects": "self"
+      }
+    ],
+    "build-base": [
+      {
+        "target": "build-base",
+        "projects": "dependencies"
+      }
+    ]
+  },
   "workspaceLayout": {
     "libsDir": "",
     "appsDir": ""
   },
   "projects": {
-    "nx": {
-      "tags": []
-    },
-    "tao": {
-      "tags": []
-    },
-    "devkit": {
-      "tags": ["tao"],
-      "implicitDependencies": ["tao"]
-    },
-    "workspace": {
-      "tags": [],
-      "implicitDependencies": ["tao", "cli", "devkit"]
-    },
-    "web": {
-      "tags": [],
-      "implicitDependencies": ["cypress", "workspace", "jest"]
-    },
-    "cypress": {
-      "tags": [],
-      "implicitDependencies": ["workspace", "linter"]
-    },
-    "jest": {
-      "tags": [],
-      "implicitDependencies": ["workspace"]
-    },
-    "react": {
-      "tags": [],
-      "implicitDependencies": ["workspace", "cypress", "jest", "web"]
-    },
-    "storybook": {
-      "tags": [],
-      "implicitDependencies": ["cypress", "workspace"]
-    },
-    "nx-plugin": {
-      "tags": [],
-      "implicitDependencies": ["workspace"]
-    },
-    "node": {
-      "tags": [],
-      "implicitDependencies": ["workspace", "jest", "linter"]
-    },
-    "next": {
-      "tags": [],
-      "implicitDependencies": ["react"]
-    },
+    "nx": {},
+    "tao": {},
+    "devkit": {},
+    "workspace": {},
+    "web": {},
+    "cypress": {},
+    "jest": {},
+    "react": {},
+    "storybook": {},
+    "nx-plugin": {},
+    "node": {},
+    "next": {},
     "nest": {
-      "tags": [],
       "implicitDependencies": ["node", "linter"]
     },
-    "linter": {
-      "tags": [],
-      "implicitDependencies": ["workspace"]
-    },
+    "linter": {},
     "express": {
-      "tags": [],
       "implicitDependencies": ["node"]
     },
-    "eslint-plugin-nx": {
-      "tags": [],
-      "implicitDependencies": ["workspace"]
-    },
+    "eslint-plugin-nx": {},
     "create-nx-workspace": {
-      "tags": [],
       "implicitDependencies": ["workspace"]
     },
     "create-nx-plugin": {
-      "tags": [],
       "implicitDependencies": ["nx-plugin"]
     },
     "cli": {
-      "tags": [],
-      "implicitDependencies": ["tao"]
+      "implicitDependencies": ["workspace"]
     },
     "angular": {
-      "tags": [],
       "implicitDependencies": ["workspace", "cypress", "jest"]
     },
-    "e2e-utils": {
-      "tags": []
-    },
+    "e2e-utils": {},
     "e2e-cli": {
-      "tags": [],
       "implicitDependencies": ["cli"]
     },
     "e2e-angular": {
-      "tags": [],
       "implicitDependencies": ["angular"]
     },
     "e2e-cypress": {
-      "tags": [],
       "implicitDependencies": ["cypress"]
     },
     "e2e-gatsby": {
-      "tags": [],
       "implicitDependencies": ["gatsby"]
     },
     "e2e-jest": {
-      "tags": [],
       "implicitDependencies": ["jest"]
     },
     "e2e-linter": {
-      "tags": [],
       "implicitDependencies": ["linter"]
     },
     "e2e-next": {
-      "tags": [],
       "implicitDependencies": ["next"]
     },
     "e2e-node": {
-      "tags": [],
       "implicitDependencies": ["node"]
     },
     "e2e-nx-plugin": {
-      "tags": [],
-      "implicitDependencies": ["nx-plugin", "create-nx-plugin"]
+      "implicitDependencies": ["create-nx-plugin"]
     },
     "e2e-react": {
-      "tags": [],
       "implicitDependencies": ["react"]
     },
     "e2e-storybook": {
-      "tags": [],
       "implicitDependencies": ["storybook"]
     },
     "e2e-web": {
-      "tags": [],
       "implicitDependencies": ["web"]
     },
     "e2e-workspace": {
-      "tags": [],
-      "implicitDependencies": ["workspace", "create-nx-workspace"]
+      "implicitDependencies": ["create-nx-workspace"]
     },
     "gatsby": {
       "tags": []
     },
     "dep-graph-dep-graph-e2e": {
-      "tags": [],
       "implicitDependencies": ["dep-graph-dep-graph"]
     },
     "dep-graph-dep-graph": {
       "tags": ["core"]
     },
-    "nx-dev": { "tags": ["scope:nx-dev", "type:app"] },
+    "nx-dev": {
+      "tags": ["scope:nx-dev", "type:app"]
+    },
     "nx-dev-e2e": {
       "tags": ["scope:nx-dev", "type:e2e"],
       "implicitDependencies": ["nx-dev"]
     },
-    "nx-dev-ui-common": { "tags": ["scope:nx-dev", "type:ui"] },
+    "nx-dev-ui-common": {
+      "tags": ["scope:nx-dev", "type:ui"]
+    },
     "nx-dev-feature-doc-viewer": {
       "tags": ["scope:nx-dev", "type:feature"]
     },
